fix(agent-runtime): extract A2aCardAutoConfiguration from RuntimeAutoConfiguration
RuntimeAutoConfiguration (boot package) should not depend on AgentCardProperties
(engine.a2a package). Extract the a2aAgentCard bean and properties registration
into a dedicated A2aCardAutoConfiguration in engine.a2a, registered via
spring auto-configuration imports.
